kvserver: move some consistency related code
Purely mechanical.

Release note: None
